Describe the configured precision correctly for mdrun
We adopted the build configuration naming scheme for GMX_DOUBLE
"mixed" vs "double" some years ago so that people knew we don't reduce
forces in mdrun solely in single. This change improves consistency in
this regard in the cases that are important to users, ie. log files
and messages about checkpoint restarts.

Various file-reading tools still refer to "single" in contexts where
they are reading floats, which is fine.