Keep the number of tMPI ranks consistent with the number of GPUs
When the number of atoms in a system is small, and the automatic resource
assignment is used (i.e., a user just runs gmx mdrun), the number of ranks
chosen by get_nthreads_mpi might be incompatible with the expectations of
GpuTaskAssignmentsBuilder, namely that the number of ranks must be
divisible by the number of threads.

If the user did not supply the -ntmpi value, and we have to limit the
number of ranks due to small system size, we round it down down to the
nearest value divisible by the number of GPUs to use (or 1, whichever is
greater).

Fixes #3874