Fix multi-sim checkpoint writing bug
In rare cases (when a checkpoint file existed on at least one multisim
rank but not on all), checkpoint writing with multiple simulation could
cause the ranks to get out of sync. This fixes the problem by moving
the barrier outside of the relevant if clause, making sure that it
is hit by all ranks.

Fixes #3919