Fix MiMiC with vsites
This moves the call to the virtual site construction in MiMiC.
The current call happened after virtual
sites are used. This call has been moved to right after the coordinates
are received. The underlying problem is likely to have lead to wrong results when using
MiMiC with virtual sites.

Refs #3866
Fixes #3904