Allow thermostat to scale velocities before and after step
Velocity scaling thermostats are currently only able to scale
velocities before the respective propagator step. This change allows
to also scale velocities at the end of the step. This functionality
is needed by the Nose-Hoover thermostat.

This requires an additional template parameter in the `updateVelocities`
function, and an additional callback in the
`PropagatorThermostatConnection` struct to allow temperature coupling
algorithms to get a view to the end velocity scaling.

This also sorts the forward declarations in modularsimulatorinterfaces.h
alphabetically to simplify follow-up changes.

Refs #3423