Use index instead of pointer in t_mde_delta_h_coll
Several members of t_mde_delta_h_coll were just pointers into a
master array of t_mde_delta_h that are held by t_mde_delta_h_coll.
This change makes it so only an int instead of a pointer is stored.
After this it should be much easier to make t_mde_delta_h_coll
default constructible, which is planned for a followup.

Also removed 2 unused members.

A renaming of the members can also be left for a followup, once
the whole t_mde_delta_h_coll struct is modernized.