Decouple update-group aspects of vsites and constraints
Update groups should not be managed from within domdec module, so
shift the responsiblity to runner temporarily to simplify future
changes.

This made clear that some internal fields of domdec module were
inaccurately named. Update groups preclude the possibility of
constraints or vsites being split across domains, but when update
groups are not in use such splits are merely possible, not
assured. The code doesn't change, because we always have to account
for the possiblity that some are split.

Refs #4004