Eliminate redundant GPU force reduction event dependency
CPU force transfers happen on the same strem as GPU reduction, so no explicit dependency is needed.

Additionally, due to a StatePropagatorDataGpu bug, whenever the force readiness event queried is not AtomLocality::All, an incorrect event is returned leading to circular dependency on force reduction, as described in #4032.

This change however does not fix the StatePropagatorDataGpu bug, but it should help avoiding workarounds for the new SYCL backend (#3932).

Refs #4032 #3988