Make ExpfitTest.EffnERREST well defined
The original test was poorly constructed, and the τ₂ fitting parameter
was not defined for the test data: with α = 1, it could take any value
and the function would fit just the same.

Here, we replace the test data with the one where 0<α<1, and thus both
τ₁ and τ₂ are well-defined.

Also updated the documentation to reflect the actual function used in
the code: it was missing a prefactor of 2.

Closes #3955