Reorganize COPYING file for auto-detection of licenses
It seems our previous complex COPYING file that did not start with
the LGPL license causes GitLab to detect our license as "other", rather
than an OSI-approved one. By clearly putting the LGPL license first
in the file and then add the more permissible exceptions we allow
at the end we avoid this.