Remove useless top level nblib force calculator
The top level nblib force calculator was a place holder in
anticipation of a unification of CPU and GPU force calculator
backends, which did not occur. As such it is removed since it only
adds maintenance burden and not utility.

Also removed is a useless class holding the non-bonded force
calculator setup director, which is now a free function.