Add constraints test with multiple molecules
In the GPU code, there is a complex logic of splitting constraints
over the thread blocks so that no coupled constraints land in
different thread blocks. To test this logic, we need the system with
more atoms than the size of the thread block. This adds such a system
to the unit test. Three atom molecule is taken as a base system since
the number of threads in a block is usually a power of 2, hence not
a multiple of 3.