Stop timers in GpuForceReduction with no atoms
The early return, as implemented earlier, was happening after starting
the timers. Now we properly stop them.

The large "if" was used here instead of just moving the "return" up in
order to accomodate future manual event handling (see Issue #3988 and MR
!2015).

Thanks to Artem Zhmurov for noticing the problem.