Improve stability of ContinuationIsExact tests
The default nstlist value in tests is set to 8. This is appropriate for
NormalIntegrators/MdrunNoAppendContinuationIsExact tests, which do the
restart after 8 steps, so the "full" trajectory and the "restarted"
trajectory update their lists at the same time. However, pairlist tuning
often increases the nstlist value, which might affect the "full" run.

Here, we force the nstlist to be half the nsteps for the "full" run by
setting it via CLI option.

Remove related known issue about Power9 test failures.

This is a backport of MR !2041 (f59f6ec6) to GROMACS 2021.

Closes #3747.