Improve trjconv test coverage
Now the essential content of the output file is checked. Saner use of
parametrized tests, along with user-friendly naming. When TNG support
is not configured, the test of TNG reading is now skipped in a way
that provides feedback about the reason for the skip.

Updated test trajectory files so that they have normal kinds of time
stamps.

Moved trajectory-reading infrastructure to testutils now that
it is used for more than mdrun tests.

Minor fixes to docs

Refs #2873