Prepare PME end-to-end tests in SetUpTestCase
As a precursor to parameterizing the mdrun end-to-pme PME tests on
their command lines, we need to shift the calls to grompp out of the
loop over those parameters. That means calling grompp during the setup
routine that precedes all of the test cases, and arranging for the
appropriate file handling for the different test flavors and rank
setups. The current implementation already did grompp early enough, so
this refactoring simply moves it earlier to clear the way for future
work.

Removed the ScaleTheBox test case, which duplicated the coverage in
the Basic one, as pbc = xyz is the default.

Refs #4255