Fix ReductionRequirement::Eventually
This reduction requirement permits subscribing modules to request that
mdrun reduces observables only when other modules are also reducing
their observables, to avoid extra simulation-wide communication.

Currently, we have most modules using compute_globals and some using
ObservablesReducer. However, the latter was not aware of when modules
using the former may have required reduction. So, the latter needs to
be told that the decision to reduce observables this step has been
made externally, so that ObservablesReducer does reduction in the case
when only ObservablesReducer subscribers that used
ReductionRequirement::Eventually have requested reduction.

This fix is needed in order that the issue of #4280 is visible also in
multi-domain runs.

Refs #4280