Remove assertion with LINCS and ObservablesReducer
This assertion isn't needed because lincs_rmsd will return zero if no
valid value is available, and the value is only used for logging
anyway. The assertion can only work when there is a one-to-one mapping
between reductions associated with any ObservablesReducer per MD step.
Currently that's not the case, because vv, vvak, and exchange methods
all use multiple calls to compute_globals every step.

Refs #4280