Use MessageStringCollector in PME end-to-end tests
This builds the skip messages and writes them to stdout, so that
later we can use them properly with GTEST_SKIP().

It also re-uses the HardwareInfo object already set up by the parent
class, rather than making a new one.

Extended functionality of CommandLine to make constructing the message
simpler to implement.

Refs #4255