Add tests for freeze groups with pressure coupling
Both Parrinello-Rahman and Berendsen barostats change coordinates of
atoms when the box is scaled. This should not affect frozen atoms,
which is now tested.

Second freeze group is reduced to side-chain hydrogens, so that there
are some unfrozen atoms in the system. Constraints are changed to
h-bonds, which is default for this test system. The tolerances are
tightened up so that the bugs are caught in these short runs.

Refs #3075