Fix pdb2gmx interactive protonation state selection
`pdb2gmx` interactive protonation state selection was not working, because the `select_res` function has switched use of `enumValueToLongString` and `enumValueToString`.

Additionally, it now shows the longer description in the prompt (`Protonated (charge 0)`) instead of just the residue name (`GLU` `GLH`).

This occurred was introduced during a modernization of gmxpreprocess to use `enum` for protonation state.