Fix missing x copy to GPU
Commit 977d0e55 fixed the stepWorkload.useGpuXBufferOps flag to
correctly be set only on non-search steps which uncovered a bug in the
code affecting cases where no force calculation requires the plain
coordinates on the GPU (e.g. without PME on search steps where x
conversion is done on the CPU). In such cases stale coordinates were
used in GPU update.

This commit is a minimal fix of the conditionals involved.

Closes #4309