Make enforced rotation group input constant
The contents of the inputrec were being changed to center reference
groups. Now copies of the inputrec are centered.

The name and docs of the field in the inputrec have been changed so
that any work on this code outside master can't silently use the old
input variable whose meaning has now changed.

Refs #3854, #3984