Note caveats when building client software in the presence of overlapping GROMACS installations.
Improve gmxapi docs for finding the installation.

The CMake config module for the gmxapi component installed with GROMACS,
does not currently support overlapping GROMACS installations, but this,
was not sufficiently clear and could lead to confusion.

In the future, we should probably make `share/cmake/gmxapi` more robust
and reduce reliance on / references to GMXRC, but these issues are
beyond the scope of the current change.

Closes #4330