Make ExponentialMovingAverage test more stable
The `ExponentialMovingAverage.DeterminesCorrectlyIfIncreasing` test was
relying on an exact equality of two floating point numbers, specifically
that 10 * (1 + (1 - 0.01) * 1) == 10 + (1 - 0.01) * 10.

To be on a safe side, we avoid testing the case of adding two exactly
same values, instead checking the behavior for a very small decrease in
the added values.

I believe that the edge-case of adding exactly the same value is not important enough, and the potential problems of relying on exact equality of FP numbers are bad enough to warrant changing the test on all architectures.

Thanks to @ElianeBriand for finding the root cause of the issue.

Closes #3196