Fix the nbnxm kernels with <8 clusters/super-cluster
A conditional was missing to ensure that i-atom prefetch does
not write out of bounds into shared memory.

This commit ensures that 4 clusters per supercluster can be enabled for
future testing.

Also added a static assert to ensure that search grid dimensions are
consistent with the number of cluster per supercluster.

Refs #3847