Run copyrights and clang-format pre merge
Move `copyright` and `clang-format` jobs from "push" to "merge" pipelines.

It's this time of the year again, we're going to update a lot of copyrights, and we will likely miss a bunch of those unless we enforce it. Running this job in "push" seems more likely to introduce annoying issues than to prevent them.

Refs #3617