Fix bug in syncing with remote coordinate availability in X GPU Halo
For multi-dimensional halo exchanges, sync on remote coordinate
availability needs to be done for every dimension, since the
remote ranks will be different for each. This fixes a bug introduced
in ccd5bef5f2, where we removed this sync for all but the first
dimension, under the false impression thatall required sync would be
satisfield implicitly.

Fixes #4351