Minor fixes to release notes
After we have merged release-2021 in today, the MR preparing for 2021.6 (not yet released) is also present in release-2022 such that the empty release notes for this possible future patch release show up on our webpage (https://manual.gromacs.org/2022-rc1/release-notes/index.html).

This MR suppresses the release notes for 2021.6 (it does not completely remove them), and moves the 2021 patch releases to their correct position on the webpage.