SYCL: Disable L0 backend with oneAPI 2022.0
There is a bug during the `sycl::queue` destruction causing either a
deadlock or an uncaught exception. It does not happen with OpenCL
backend or with older oneAPI versions.

In mdrun, the bug only occurs after the simulation ends, not affecting
the data but causing either a time-out or a non-zero return code.

In the interest of preventing user error, we mark LevelZero devices
as incompatible when using oneAPI 2021.0.

The issue also affects open-source IntelLLVM, which has a dynamic
version value. We don't handle this case and trust that users who
use nightly compiler builds don't expect stability anyway.

Refs #4354